Split ProfileMk3 node implementation into several modules.
