+++ nodes/generators_extended/profile_mk3.py
@@ -16,25 +16,19 @@
 #
 # ##### END GPL LICENSE BLOCK #####
 
-import re
-from itertools import chain
-import ast
-from math import *
 import os
 
 import bpy
 from bpy.props import BoolProperty, StringProperty, EnumProperty, FloatProperty, IntProperty
-from mathutils.geometry import interpolate_bezier
-from mathutils import Vector, Matrix
 
 from sverchok.node_tree import SverchCustomTreeNode
 from sverchok.core.socket_data import SvNoDataError
-from sverchok.data_structure import fullList, updateNode, dataCorrect, match_long_repeat
-from sverchok.utils.parsec import *
+from sverchok.data_structure import updateNode, match_long_repeat
 from sverchok.utils.logging import info, debug, warning
-from sverchok.utils.geom import interpolate_quadratic_bezier
-from sverchok.utils.sv_curve_utils import Arc
 from sverchok.utils.sv_update_utils import sv_get_local_path
+
+from sverchok.utils.modules.profile_mk3.interpreter import Interpreter
+from sverchok.utils.modules.profile_mk3.parser import parse_profile
 
 '''
 input like:
@@ -121,1085 +115,6 @@ Our DSL has relatively simple BNF:
     <Const> ::= Python integer or floating-point literal
 
 """
-
-def make_functions_dict(*functions):
-    return dict([(function.__name__, function) for function in functions])
-
-# Functions
-safe_names = make_functions_dict(
-        # From math module
-        acos, acosh, asin, asinh, atan, atan2,
-        atanh, ceil, copysign, cos, cosh, degrees,
-        erf, erfc, exp, expm1, fabs, factorial, floor,
-        fmod, frexp, fsum, gamma, hypot, isfinite, isinf,
-        isnan, ldexp, lgamma, log, log10, log1p, log2, modf,
-        pow, radians, sin, sinh, sqrt, tan, tanh, trunc,
-        # Additional functions
-        abs,
-        # From mathutlis module
-        Vector, Matrix,
-        # Python type conversions
-        tuple, list, str
-    )
-# Constants
-safe_names['e'] = e
-safe_names['pi'] = pi
-
-##########################################
-# Expression classes
-##########################################
-
-class Expression(object):
-    def __init__(self, expr, string):
-        self.expr = expr
-        self.string = string
-
-
-    def __repr__(self):
-        return "Expr({})".format(self.string)
-
-    @classmethod
-    def from_string(cls, string):
-        try:
-            string = string[1:][:-1]
-            expr = ast.parse(string, mode='eval')
-            return Expression(expr, string)
-        except Exception as e:
-            print(e)
-            print(string)
-            return None
-
-    def eval_(self, variables):
-        env = dict()
-        env.update(safe_names)
-        env.update(variables)
-        env["__builtins__"] = {}
-        return eval(compile(self.expr, "<expression>", 'eval'), env)
-
-    def get_variables(self):
-        result = {node.id for node in ast.walk(self.expr) if isinstance(node, ast.Name)}
-        return result.difference(safe_names.keys())
-
-class Const(Expression):
-    def __init__(self, value):
-        self.value = value
-
-    @classmethod
-    def from_string(cls, string):
-        try:
-            return Const( float(string) )
-        except ValueError:
-            return None
-
-    def __repr__(self):
-        return "Const({})".format(self.value)
-
-    def eval_(self, variables):
-        return self.value
-
-    def get_variables(self):
-        return set()
-
-class Variable(Expression):
-    def __init__(self, name):
-        self.name = name
-
-    @classmethod
-    def from_string(cls, string):
-        return Variable(string)
-
-    def __repr__(self):
-        return "Variable({})".format(self.name)
-
-    def eval_(self, variables):
-        value = variables.get(self.name, None)
-        if value is not None:
-            return value
-        else:
-            raise SyntaxError("Unknown variable: " + self.name)
-
-    def get_variables(self):
-        return set([self.name])
-
-# In general, this does not have very much sense:
-# instead of -a one can write {-a}, then it will
-# be parsed as Expression and will work fine.
-# This is mostly implemented for compatibility
-# with older Profile node syntax.
-class NegatedVariable(Variable):
-    @classmethod
-    def from_string(cls, string):
-        return NegatedVariable(string)
-
-    def eval_(self, variables):
-        value = variables.get(self.name, None)
-        if value is not None:
-            return -value
-        else:
-            raise SyntaxError("Unknown variable: " + self.name)
-
-    def __repr__(self):
-        return "NegatedVariable({})".format(self.name)
-
-############################################
-# Statement classes
-# Classes for AST of our DSL
-############################################
-
-# These classes are responsible for interpretation of specific DSL statements.
-# Each of these classes does the following:
-# 
-#  * Stores statement parameters (for example, MoveTo stores x and y).
-#  * defines get_variables() method, which should return a set of all
-#    variables used by all expressions in the statement.
-#  * defines interpret() method, which should calculate all vertices and
-#    edges according to statement parameters, and pass them to the interpreter.
-
-class Statement(object):
-    
-    def get_variables(self):
-        return set()
-
-    def get_hidden_inputs(self):
-        return set()
-
-    def get_optional_inputs(self):
-        return set()
-
-class MoveTo(Statement):
-    def __init__(self, is_abs, x, y):
-        self.is_abs = is_abs
-        self.x = x
-        self.y = y
-
-    def __repr__(self):
-        letter = "M" if self.is_abs else "m"
-        return "{} {} {}".format(letter, self.x, self.y)
-
-    def get_variables(self):
-        variables = set()
-        variables.update(self.x.get_variables())
-        variables.update(self.y.get_variables())
-        return variables
-
-    def interpret(self, interpreter, variables):
-        interpreter.assert_not_closed()
-        interpreter.start_new_segment()
-        pos = interpreter.calc_vertex(self.is_abs, self.x, self.y, variables)
-        interpreter.position = pos
-        interpreter.new_knot("M.#", *pos)
-        interpreter.has_last_vertex = False
-
-class LineTo(Statement):
-    def __init__(self, is_abs, pairs, close):
-        self.is_abs = is_abs
-        self.pairs = pairs
-        self.close = close
-
-    def get_variables(self):
-        variables = set()
-        for x, y in self.pairs:
-            variables.update(x.get_variables())
-            variables.update(y.get_variables())
-        return variables
-
-    def __repr__(self):
-        letter = "L" if self.is_abs else "l"
-        return "{} {} {}".format(letter, self.pairs, self.close)
-    
-    def interpret(self, interpreter, variables):
-        interpreter.assert_not_closed()
-        interpreter.start_new_segment()
-        v0 = interpreter.position
-        if interpreter.has_last_vertex:
-            v0_index = interpreter.get_last_vertex()
-        else:
-            v0_index = interpreter.new_vertex(*v0)
-
-        for i, (x_expr, y_expr) in enumerate(self.pairs):
-            v1 = interpreter.calc_vertex(self.is_abs, x_expr, y_expr, variables)
-            interpreter.position = v1
-            v1_index = interpreter.new_vertex(*v1)
-            interpreter.new_edge(v0_index, v1_index)
-            interpreter.new_knot("L#.{}".format(i), *v1)
-            v0_index = v1_index
-
-        if self.close:
-            interpreter.new_edge(v1_index, interpreter.segment_start_index)
-
-        interpreter.has_last_vertex = True
-
-class HorizontalLineTo(Statement):
-    def __init__(self, is_abs, xs):
-        self.is_abs = is_abs
-        self.xs = xs
-
-    def get_variables(self):
-        variables = set()
-        for x in self.xs:
-            variables.update(x.get_variables())
-        return variables
-
-    def __repr__(self):
-        letter = "H" if self.is_abs else "h"
-        return "{} {}".format(letter, self.xs)
-
-    def interpret(self, interpreter, variables):
-        interpreter.assert_not_closed()
-        interpreter.start_new_segment()
-        v0 = interpreter.position
-        if interpreter.has_last_vertex:
-            v0_index = interpreter.get_last_vertex()
-        else:
-            v0_index = interpreter.new_vertex(*v0)
-
-        for i, x_expr in enumerate(self.xs):
-            x0,y0 = interpreter.position
-            x = interpreter.eval_(x_expr, variables)
-            if not self.is_abs:
-                x = x0 + x
-            v1 = (x, y0)
-            interpreter.position = v1
-            v1_index = interpreter.new_vertex(*v1)
-            interpreter.new_edge(v0_index, v1_index)
-            interpreter.new_knot("H#.{}".format(i), *v1)
-            v0_index = v1_index
-
-        interpreter.has_last_vertex = True
-
-class VerticalLineTo(Statement):
-    def __init__(self, is_abs, ys):
-        self.is_abs = is_abs
-        self.ys = ys
-
-    def get_variables(self):
-        variables = set()
-        for y in self.ys:
-            variables.update(y.get_variables())
-        return variables
-
-    def __repr__(self):
-        letter = "V" if self.is_abs else "v"
-        return "{} {}".format(letter, self.ys)
-
-    def interpret(self, interpreter, variables):
-        interpreter.assert_not_closed()
-        interpreter.start_new_segment()
-        v0 = interpreter.position
-        if interpreter.has_last_vertex:
-            v0_index = interpreter.get_last_vertex()
-        else:
-            v0_index = interpreter.new_vertex(*v0)
-
-        for i, y_expr in enumerate(self.ys):
-            x0,y0 = interpreter.position
-            y = interpreter.eval_(y_expr, variables)
-            if not self.is_abs:
-                y = y0 + y
-            v1 = (x0, y)
-            interpreter.position = v1
-            v1_index = interpreter.new_vertex(*v1)
-            interpreter.new_edge(v0_index, v1_index)
-            interpreter.new_knot("V#.{}".format(i), *v1)
-            v0_index = v1_index
-
-        interpreter.has_last_vertex = True
-
-class CurveTo(Statement):
-    class Segment(object):
-        def __init__(self, control1, control2, knot2):
-            self.control1 = control1
-            self.control2 = control2
-            self.knot2 = knot2
-
-        def __repr__(self):
-            return "{} {} {}".format(self.control1, self.control2, self.knot2)
-
-    def __init__(self, is_abs, segments, num_segments, close):
-        self.is_abs = is_abs
-        self.segments = segments
-        self.num_segments = num_segments
-        self.close = close
-
-    def get_variables(self):
-        variables = set()
-        for segment in self.segments:
-            variables.update(segment.control1[0].get_variables())
-            variables.update(segment.control1[1].get_variables())
-            variables.update(segment.control2[0].get_variables())
-            variables.update(segment.control2[1].get_variables())
-            variables.update(segment.knot2[0].get_variables())
-            variables.update(segment.knot2[1].get_variables())
-        if self.num_segments:
-            variables.update(self.num_segments.get_variables())
-        return variables
-
-    def __repr__(self):
-        letter = "C" if self.is_abs else "c"
-        segments = " ".join(str(segment) for segment in self.segments)
-        return "{} {} n={} {}".format(letter, segments, self.num_segments, self.close)
-
-    def interpret(self, interpreter, variables):
-        vec = lambda v: Vector((v[0], v[1], 0))
-
-        interpreter.assert_not_closed()
-        interpreter.start_new_segment()
-
-        v0 = interpreter.position
-        if interpreter.has_last_vertex:
-            v0_index = interpreter.get_last_vertex()
-        else:
-            v0_index = interpreter.new_vertex(*v0)
-
-        knot1 = None
-        for i, segment in enumerate(self.segments):
-            # For first segment, knot1 is initial pen position;
-            # for the following, knot1 is knot2 of previous segment.
-            if knot1 is None:
-                knot1 = interpreter.position
-            else:
-                knot1 = knot2
-
-            handle1 = interpreter.calc_vertex(self.is_abs, segment.control1[0], segment.control1[1], variables)
-
-            # In Profile mk2, for "c" handle2 was calculated relative to handle1,
-            # and knot2 was calculated relative to handle2.
-            # But in SVG specification, 
-            # >> ...  *At the end of the command*, the new current point becomes
-            # >> the final (x,y) coordinate pair used in the polybézier.
-            # This is also behaivour of browsers.
-
-            #interpreter.position = handle1
-            handle2 = interpreter.calc_vertex(self.is_abs, segment.control2[0], segment.control2[1], variables)
-            #interpreter.position = handle2
-            knot2 = interpreter.calc_vertex(self.is_abs, segment.knot2[0], segment.knot2[1], variables)
-
-            if self.num_segments is not None:
-                r = interpreter.eval_(self.num_segments, variables)
-            else:
-                r = interpreter.dflt_num_verts
-
-            points = interpolate_bezier(vec(knot1), vec(handle1), vec(handle2), vec(knot2), r)
-
-            interpreter.new_knot("C#.{}.h1".format(i), *handle1)
-            interpreter.new_knot("C#.{}.h2".format(i), *handle2)
-            interpreter.new_knot("C#.{}.k".format(i), *knot2)
-
-            interpreter.prev_bezier_knot = handle2
-
-            for point in points[1:]:
-                v1_index = interpreter.new_vertex(point.x, point.y)
-                interpreter.new_edge(v0_index, v1_index)
-                v0_index = v1_index
-
-        interpreter.position = knot2
-
-        if self.close:
-            interpreter.new_edge(v1_index, interpreter.segment_start_index)
-
-        interpreter.has_last_vertex = True
-
-class SmoothCurveTo(Statement):
-    class Segment(object):
-        def __init__(self, control2, knot2):
-            self.control2 = control2
-            self.knot2 = knot2
-
-        def __repr__(self):
-            return "{} {}".format(self.control2, self.knot2)
-
-    def __init__(self, is_abs, segments, num_segments, close):
-        self.is_abs = is_abs
-        self.segments = segments
-        self.num_segments = num_segments
-        self.close = close
-
-    def get_variables(self):
-        variables = set()
-        for segment in self.segments:
-            variables.update(segment.control2[0].get_variables())
-            variables.update(segment.control2[1].get_variables())
-            variables.update(segment.knot2[0].get_variables())
-            variables.update(segment.knot2[1].get_variables())
-        if self.num_segments:
-            variables.update(self.num_segments.get_variables())
-        return variables
-
-    def __repr__(self):
-        letter = "S" if self.is_abs else "s"
-        segments = " ".join(str(segment) for segment in self.segments)
-        return "{} {} n={} {}".format(letter, segments, self.num_segments, self.close)
-
-    def interpret(self, interpreter, variables):
-        vec = lambda v: Vector((v[0], v[1], 0))
-
-        interpreter.assert_not_closed()
-        interpreter.start_new_segment()
-
-        v0 = interpreter.position
-        if interpreter.has_last_vertex:
-            v0_index = interpreter.get_last_vertex()
-        else:
-            v0_index = interpreter.new_vertex(*v0)
-
-        knot1 = None
-        for i, segment in enumerate(self.segments):
-            # For first segment, knot1 is initial pen position;
-            # for the following, knot1 is knot2 of previous segment.
-            if knot1 is None:
-                knot1 = interpreter.position
-            else:
-                knot1 = knot2
-
-            if interpreter.prev_bezier_knot is None:
-                # If there is no previous command or if the previous command was
-                # not an C, c, S or s, assume the first control point is coincident
-                # with the current point.
-                handle1 = knot1
-            else:
-                # The first control point is assumed to be the reflection of the
-                # second control point on the previous command relative to the
-                # current point. 
-                prev_knot_x, prev_knot_y = interpreter.prev_bezier_knot
-                x0, y0 = knot1
-                dx, dy = x0 - prev_knot_x, y0 - prev_knot_y
-                handle1 = x0 + dx, y0 + dy
-
-            # I assume that handle2 should be relative to knot1, not to handle1.
-            # interpreter.position = handle1
-            handle2 = interpreter.calc_vertex(self.is_abs, segment.control2[0], segment.control2[1], variables)
-            # interpreter.position = handle2
-            knot2 = interpreter.calc_vertex(self.is_abs, segment.knot2[0], segment.knot2[1], variables)
-
-            if self.num_segments is not None:
-                r = interpreter.eval_(self.num_segments, variables)
-            else:
-                r = interpreter.dflt_num_verts
-
-            points = interpolate_bezier(vec(knot1), vec(handle1), vec(handle2), vec(knot2), r)
-
-            interpreter.new_knot("S#.{}.h1".format(i), *handle1)
-            interpreter.new_knot("S#.{}.h2".format(i), *handle2)
-            interpreter.new_knot("S#.{}.k".format(i), *knot2)
-
-            interpreter.prev_bezier_knot = handle2
-
-            for point in points[1:]:
-                v1_index = interpreter.new_vertex(point.x, point.y)
-                interpreter.new_edge(v0_index, v1_index)
-                v0_index = v1_index
-
-        interpreter.position = knot2
-
-        if self.close:
-            interpreter.new_edge(v1_index, interpreter.segment_start_index)
-
-        interpreter.has_last_vertex = True
-
-class QuadraticCurveTo(Statement):
-    class Segment(object):
-        def __init__(self, control, knot2):
-            self.control = control
-            self.knot2 = knot2
-
-        def __repr__(self):
-            return "{} {}".format(self.control, self.knot2)
-
-    def __init__(self, is_abs, segments, num_segments, close):
-        self.is_abs = is_abs
-        self.segments = segments
-        self.num_segments = num_segments
-        self.close = close
-
-    def get_variables(self):
-        variables = set()
-        for segment in self.segments:
-            variables.update(segment.control[0].get_variables())
-            variables.update(segment.control[1].get_variables())
-            variables.update(segment.knot2[0].get_variables())
-            variables.update(segment.knot2[1].get_variables())
-        if self.num_segments:
-            variables.update(self.num_segments.get_variables())
-        return variables
-
-    def __repr__(self):
-        letter = "Q" if self.is_abs else "q"
-        segments = " ".join(str(segment) for segment in self.segments)
-        return "{} {} n={} {}".format(letter, segments, self.num_segments, self.close)
-
-    def interpret(self, interpreter, variables):
-        vec = lambda v: Vector((v[0], v[1], 0))
-
-        interpreter.assert_not_closed()
-        interpreter.start_new_segment()
-
-        v0 = interpreter.position
-        if interpreter.has_last_vertex:
-            v0_index = interpreter.get_last_vertex()
-        else:
-            v0_index = interpreter.new_vertex(*v0)
-
-        knot1 = None
-        for i, segment in enumerate(self.segments):
-            # For first segment, knot1 is initial pen position;
-            # for the following, knot1 is knot2 of previous segment.
-            if knot1 is None:
-                knot1 = interpreter.position
-            else:
-                knot1 = knot2
-
-            handle = interpreter.calc_vertex(self.is_abs, segment.control[0], segment.control[1], variables)
-            knot2 = interpreter.calc_vertex(self.is_abs, segment.knot2[0], segment.knot2[1], variables)
-
-            if self.num_segments is not None:
-                r = interpreter.eval_(self.num_segments, variables)
-            else:
-                r = interpreter.dflt_num_verts
-
-            points = interpolate_quadratic_bezier(vec(knot1), vec(handle), vec(knot2), r)
-
-            interpreter.new_knot("Q#.{}.h".format(i), *handle)
-            interpreter.new_knot("Q#.{}.k".format(i), *knot2)
-
-            interpreter.prev_quad_bezier_knot = handle
-
-            for point in points[1:]:
-                v1_index = interpreter.new_vertex(point.x, point.y)
-                interpreter.new_edge(v0_index, v1_index)
-                v0_index = v1_index
-
-        interpreter.position = knot2
-
-        if self.close:
-            interpreter.new_edge(v1_index, interpreter.segment_start_index)
-
-        interpreter.has_last_vertex = True
-
-class SmoothQuadraticCurveTo(Statement):
-    class Segment(object):
-        def __init__(self, knot2):
-            self.knot2 = knot2
-
-        def __repr__(self):
-            return str(self.knot2)
-
-    def __init__(self, is_abs, segments, num_segments, close):
-        self.is_abs = is_abs
-        self.segments = segments
-        self.num_segments = num_segments
-        self.close = close
-
-    def get_variables(self):
-        variables = set()
-        for segment in self.segments:
-            variables.update(segment.knot2[0].get_variables())
-            variables.update(segment.knot2[1].get_variables())
-        if self.num_segments:
-            variables.update(self.num_segments.get_variables())
-        return variables
-
-    def __repr__(self):
-        letter = "T" if self.is_abs else "t"
-        segments = " ".join(str(segment) for segment in self.segments)
-        return "{} {} n={} {}".format(letter, segments, self.num_segments, self.close)
-
-    def interpret(self, interpreter, variables):
-        vec = lambda v: Vector((v[0], v[1], 0))
-
-        interpreter.assert_not_closed()
-        interpreter.start_new_segment()
-
-        v0 = interpreter.position
-        if interpreter.has_last_vertex:
-            v0_index = interpreter.get_last_vertex()
-        else:
-            v0_index = interpreter.new_vertex(*v0)
-
-        knot1 = None
-        for i, segment in enumerate(self.segments):
-            # For first segment, knot1 is initial pen position;
-            # for the following, knot1 is knot2 of previous segment.
-            if knot1 is None:
-                knot1 = interpreter.position
-            else:
-                knot1 = knot2
-
-            if interpreter.prev_quad_bezier_knot is None:
-                # If there is no previous command or if the previous command was
-                # not a Q, q, T or t, assume the control point is coincident with
-                # the current point.
-                handle = knot1
-            else:
-                # The first control point is assumed to be the reflection of the
-                # second control point on the previous command relative to the
-                # current point. 
-                prev_knot_x, prev_knot_y = interpreter.prev_quad_bezier_knot
-                x0, y0 = knot1
-                dx, dy = x0 - prev_knot_x, y0 - prev_knot_y
-                handle = x0 + dx, y0 + dy
-
-            knot2 = interpreter.calc_vertex(self.is_abs, segment.knot2[0], segment.knot2[1], variables)
-
-            if self.num_segments is not None:
-                r = interpreter.eval_(self.num_segments, variables)
-            else:
-                r = interpreter.dflt_num_verts
-
-            points = interpolate_quadratic_bezier(vec(knot1), vec(handle), vec(knot2), r)
-
-            interpreter.new_knot("T#.{}.h".format(i), *handle)
-            interpreter.new_knot("T#.{}.k".format(i), *knot2)
-
-            interpreter.prev_quad_bezier_knot = handle
-
-            for point in points[1:]:
-                v1_index = interpreter.new_vertex(point.x, point.y)
-                interpreter.new_edge(v0_index, v1_index)
-                v0_index = v1_index
-
-        interpreter.position = knot2
-
-        if self.close:
-            interpreter.new_edge(v1_index, interpreter.segment_start_index)
-
-        interpreter.has_last_vertex = True
-
-class ArcTo(Statement):
-    def __init__(self, is_abs, radii, rot, flag1, flag2, end, num_verts, close):
-        self.is_abs = is_abs
-        self.radii = radii
-        self.rot = rot
-        self.flag1 = flag1
-        self.flag2 = flag2
-        self.end = end
-        self.num_verts = num_verts
-        self.close = close
-
-    def get_variables(self):
-        variables = set()
-        variables.update(self.radii[0].get_variables())
-        variables.update(self.radii[1].get_variables())
-        variables.update(self.rot.get_variables())
-        variables.update(self.flag1.get_variables())
-        variables.update(self.flag2.get_variables())
-        variables.update(self.end[0].get_variables())
-        variables.update(self.end[1].get_variables())
-        if self.num_verts:
-            variables.update(self.num_verts.get_variables())
-        return variables
-
-    def __repr__(self):
-        letter = "A" if self.is_abs else "a"
-        return "{} {} {} {} {} {} n={} {}".format(letter, self.radii, self.rot, self.flag1, self.flag2, self.end, self.num_verts, self.close)
-
-    def interpret(self, interpreter, variables):
-        interpreter.assert_not_closed()
-        interpreter.start_new_segment()
-
-        v0 = interpreter.position
-        if interpreter.has_last_vertex:
-            v0_index = interpreter.get_last_vertex()
-        else:
-            v0_index = interpreter.new_vertex(*v0)
-
-        start = complex(*v0)
-        rad_x_expr, rad_y_expr = self.radii
-        rad_x = interpreter.eval_(rad_x_expr, variables)
-        rad_y = interpreter.eval_(rad_y_expr, variables)
-        radius = complex(rad_x, rad_y)
-        xaxis_rot = interpreter.eval_(self.rot, variables)
-        flag1 = interpreter.eval_(self.flag1, variables)
-        flag2 = interpreter.eval_(self.flag2, variables)
-
-        # numverts, requires -1 else it means segments (21 verts is 20 segments).
-        if self.num_verts is not None:
-            num_verts = interpreter.eval_(self.num_verts, variables)
-        else:
-            num_verts = interpreter.dflt_num_verts
-        num_verts -= 1
-
-        end = interpreter.calc_vertex(self.is_abs, self.end[0], self.end[1], variables)
-        end = complex(*end)
-
-        arc = Arc(start, radius, xaxis_rot, flag1, flag2, end)
-
-        theta = 1/num_verts
-        for i in range(num_verts+1):
-            v1 = x, y = arc.point(theta * i)
-            v1_index = interpreter.new_vertex(x, y)
-            interpreter.new_edge(v0_index, v1_index)
-            v0_index = v1_index
-
-        interpreter.position = v1
-        interpreter.new_knot("A.#", *v1)
-        if self.close:
-            interpreter.new_edge(v1_index, interpreter.segment_start_index)
-
-        interpreter.has_last_vertex = True
-
-class Close(Statement):
-    def __init__(self):
-        pass
-
-    def __repr__(self):
-        return "X"
-
-    def get_variables(self):
-        return set()
-
-    def interpret(self, interpreter, variables):
-        interpreter.assert_not_closed()
-        if not interpreter.has_last_vertex:
-            info("X statement: no current point, do nothing")
-            return
-
-        v0 = interpreter.vertices[0]
-        v1 = interpreter.vertices[-1]
-
-        distance = (Vector(v0) - Vector(v1)).length
-
-        if distance < interpreter.close_threshold:
-            interpreter.pop_last_vertex()
-
-        v1_index = interpreter.get_last_vertex()
-        interpreter.new_edge(v1_index, 0)
-        interpreter.closed = True
-
-class Default(Statement):
-    def __init__(self, name, value):
-        self.name = name
-        self.value = value
-
-    def __repr__(self):
-        return "default {} = {}".format(self.name, self.value)
-
-    def get_variables(self):
-        return self.value.get_variables()
-
-    def get_optional_inputs(self):
-        return set([self.name])
-
-    def interpret(self, interpreter, variables):
-        if self.name in interpreter.defaults:
-            raise Exception("Value for the `{}' variable has been already assigned!".format(self.name))
-        if self.name not in interpreter.input_names:
-            value = interpreter.eval_(self.value, variables)
-            interpreter.defaults[self.name] = value
-
-class Assign(Default):
-    def __repr__(self):
-        return "let {} = {}".format(self.name, self.value)
-
-    def get_hidden_inputs(self):
-        return set([self.name])
-
-#########################################
-# DSL parsing
-#########################################
-
-# Compare these definitions with BNF definition at the top
-
-expr_regex = re.compile(r"({[^}]+})\s*", re.DOTALL)
-
-def parse_expr(src):
-    for string, rest in parse_regexp(expr_regex)(src):
-        expr = Expression.from_string(string)
-        if expr is not None:
-            yield expr, rest
-
-identifier_regexp = re.compile(r"[a-zA-Z_][a-zA-Z0-9_]*")
-
-parse_semicolon = parse_word(";")
-
-def parse_identifier(src):
-    for (name, _), rest in sequence(parse_regexp(identifier_regexp), parse_whitespace)(src):
-        yield name, rest
-
-def parse_negated_variable(src):
-    for (_, name, _), rest in sequence(parse_word("-"), parse_regexp(identifier_regexp), parse_whitespace)(src):
-        yield NegatedVariable(name), rest
-
-def parse_value(src):
-    for smth, rest in one_of(parse_number, parse_identifier, parse_negated_variable, parse_expr)(src):
-        if isinstance(smth, (int, float)):
-            yield Const(smth), rest
-        elif isinstance(smth, str):
-            yield Variable(smth), rest
-        else:
-            yield smth, rest
-
-def parse_pair(src):
-    parser = sequence(parse_value, parse_word(","), parse_value)
-    for (x, _, y), rest in parser(src):
-        yield (x,y), rest
-
-def parse_letter(absolute, relative):
-    def parser(src):
-        for smth, rest in one_of(parse_word(absolute), parse_word(relative))(src):
-            is_abs = smth == absolute
-            yield is_abs, rest
-    return parser
-
-def parse_MoveTo(src):
-    parser = sequence(parse_letter("M", "m"), parse_pair, optional(parse_semicolon))
-    for (is_abs, (x, y), _), rest in parser(src):
-        yield MoveTo(is_abs, x, y), rest
-
-def parse_LineTo(src):
-    parser = sequence(
-                parse_letter("L", "l"),
-                many(parse_pair),
-                optional(parse_word("z")),
-                optional(parse_semicolon))
-    for (is_abs, pairs, z, _), rest in parser(src):
-        yield LineTo(is_abs, pairs, z is not None), rest
-
-def parse_parameter(name):
-    def parser(src):
-        for (_, _, value), rest in sequence(parse_word(name), parse_word("="), parse_value)(src):
-            yield value, rest
-    return parser
-
-def parse_CurveTo(src):
-
-    def parse_segment(src):
-        parser = sequence(parse_pair, parse_pair, parse_pair)
-        for (control1, control2, knot2), rest in parser(src):
-            yield CurveTo.Segment(control1, control2, knot2), rest
-
-    parser = sequence(
-                parse_letter("C", "c"),
-                many(parse_segment),
-                optional(parse_parameter("n")),
-                optional(parse_word("z")),
-                optional(parse_semicolon)
-            )
-    for (is_abs, segments, num_segments, z, _), rest in parser(src):
-        yield CurveTo(is_abs, segments, num_segments, z is not None), rest
-
-def parse_SmoothCurveTo(src):
-
-    def parse_segment(src):
-        parser = sequence(parse_pair, parse_pair)
-        for (control2, knot2), rest in parser(src):
-            yield SmoothCurveTo.Segment(control2, knot2), rest
-
-    parser = sequence(
-                parse_letter("S", "s"),
-                many(parse_segment),
-                optional(parse_parameter("n")),
-                optional(parse_word("z")),
-                optional(parse_semicolon)
-            )
-    for (is_abs, segments, num_segments, z, _), rest in parser(src):
-        yield SmoothCurveTo(is_abs, segments, num_segments, z is not None), rest
-
-def parse_QuadCurveTo(src):
-
-    def parse_segment(src):
-        parser = sequence(parse_pair, parse_pair)
-        for (control, knot2), rest in parser(src):
-            yield QuadraticCurveTo.Segment(control, knot2), rest
-
-    parser = sequence(
-                parse_letter("Q", "q"),
-                many(parse_segment),
-                optional(parse_parameter("n")),
-                optional(parse_word("z")),
-                optional(parse_semicolon)
-            )
-    for (is_abs, segments, num_segments, z, _), rest in parser(src):
-        yield QuadraticCurveTo(is_abs, segments, num_segments, z is not None), rest
-
-def parse_SmoothQuadCurveTo(src):
-
-    def parse_segment(src):
-        for knot2, rest in parse_pair(src):
-            yield SmoothQuadraticCurveTo.Segment(knot2), rest
-
-    parser = sequence(
-                parse_letter("T", "t"),
-                many(parse_segment),
-                optional(parse_parameter("n")),
-                optional(parse_word("z")),
-                optional(parse_semicolon)
-            )
-    for (is_abs, knot2, num_segments, z, _), rest in parser(src):
-        yield SmoothQuadraticCurveTo(is_abs, knot2, num_segments, z is not None), rest
-
-def parse_ArcTo(src):
-    parser = sequence(
-                parse_letter("A", "a"),
-                parse_pair,
-                parse_value,
-                parse_value,
-                parse_value,
-                parse_pair,
-                optional(parse_parameter("n")),
-                optional(parse_word("z")),
-                optional(parse_semicolon)
-            )
-    for (is_abs, radii, rot, flag1, flag2, end, num_verts, z, _), rest in parser(src):
-        yield ArcTo(is_abs, radii, rot, flag1, flag2, end, num_verts, z is not None), rest
-
-def parse_HorLineTo(src):
-    # NB: H/h command MUST end with semicolon, otherwise we will not be able to
-    # understand where it ends, i.e. does the following letter begin a new statement
-    # or is it just next X value denoted by variable.
-    parser = sequence(parse_letter("H", "h"), many(parse_value), parse_semicolon)
-    for (is_abs, xs, _), rest in parser(src):
-        yield HorizontalLineTo(is_abs, xs), rest
-
-def parse_VertLineTo(src):
-    # NB: V/v command MUST end with semicolon, otherwise we will not be able to
-    # understand where it ends, i.e. does the following letter begin a new statement
-    # or is it just next X value denoted by variable.
-    parser = sequence(parse_letter("V", "v"), many(parse_value), parse_semicolon)
-    for (is_abs, ys, _), rest in parser(src):
-        yield VerticalLineTo(is_abs, ys), rest
-
-parse_Close = parse_word("X", Close())
-
-def parse_Default(src):
-    parser = sequence(
-                parse_word("default"),
-                parse_identifier,
-                parse_word("="),
-                parse_value,
-                optional(parse_semicolon)
-            )
-    for (_, name, _, value, _), rest in parser(src):
-        yield Default(name, value), rest
-
-def parse_Assign(src):
-    parser = sequence(
-                parse_word("let"),
-                parse_identifier,
-                parse_word("="),
-                parse_value,
-                optional(parse_semicolon)
-            )
-    for (_, name, _, value, _), rest in parser(src):
-        yield Assign(name, value), rest
-
-parse_statement = one_of(
-                    parse_Default,
-                    parse_Assign,
-                    parse_MoveTo,
-                    parse_LineTo,
-                    parse_HorLineTo,
-                    parse_VertLineTo,
-                    parse_CurveTo,
-                    parse_SmoothCurveTo,
-                    parse_QuadCurveTo,
-                    parse_SmoothQuadCurveTo,
-                    parse_ArcTo,
-                    parse_Close)
-
-parse_definition = many(parse_statement)
-
-def parse_profile(src):
-    # Strip comments
-    # (hope noone uses # in expressions)
-    cleaned = ""
-    for line in src.split("\n"):
-        comment_idx = line.find('#')
-        if comment_idx != -1:
-            line = line[:comment_idx]
-        cleaned = cleaned + " " + line
-    
-    profile = parse(parse_definition, cleaned)
-    debug(profile)
-    return profile
-
-#################################
-# DSL Interpreter
-#################################
-
-# This class does the following:
-#
-# * Stores the "drawing" state, such as "current pen position"
-# * Provides API for Statement classes to add vertices, edges to the current
-#   drawing
-# * Contains the interpret() method, which runs the whole interpretation process.
-
-class Interpreter(object):
-    def __init__(self, node, input_names):
-        self.position = (0, 0)
-        self.next_vertex_index = 0
-        self.segment_start_index = 0
-        self.segment_number = 0
-        self.has_last_vertex = False
-        self.closed = False
-        self.prev_bezier_knot = None
-        self.prev_quad_bezier_knot = None
-        self.vertices = []
-        self.edges = []
-        self.knots = []
-        self.knotnames = []
-        self.dflt_num_verts = node.curve_points_count
-        self.close_threshold = node.close_threshold
-        self.defaults = dict()
-        self.input_names = input_names
-
-    def assert_not_closed(self):
-        if self.closed:
-            raise Exception("Path was already closed, will not process any further directives!")
-
-    def relative(self, x, y):
-        x0, y0 = self.position
-        return x0+x, y0+y
-
-    def calc_vertex(self, is_abs, x_expr, y_expr, variables):
-        x = self.eval_(x_expr, variables)
-        y = self.eval_(y_expr, variables)
-        if is_abs:
-            return x,y
-        else:
-            return self.relative(x,y)
-
-    def new_vertex(self, x, y):
-        index = self.next_vertex_index
-        vertex = (x, y)
-        self.vertices.append(vertex)
-        self.next_vertex_index += 1
-        return index
-
-    def new_edge(self, v1, v2):
-        self.edges.append((v1, v2))
-
-    def new_knot(self, name, x, y):
-        self.knots.append((x, y))
-        name = name.replace("#", str(self.segment_number))
-        self.knotnames.append(name)
-
-    def start_new_segment(self):
-        self.segment_start_index = self.next_vertex_index
-        self.segment_number += 1
-
-    def get_last_vertex(self):
-        return self.next_vertex_index - 1
-
-    def pop_last_vertex(self):
-        self.vertices.pop()
-        self.next_vertex_index -= 1
-        is_not_last = lambda e: e[0] != self.next_vertex_index and e[1] != self.next_vertex_index
-        self.edges = list(filter(is_not_last, self.edges))
-
-    def eval_(self, expr, variables):
-        variables_ = self.defaults.copy()
-        for name in variables:
-            value = variables[name]
-            if value is not None:
-                variables_[name] = value
-        return expr.eval_(variables_)
-
-    def interpret(self, profile, variables):
-        if not profile:
-            return
-        for statement in profile:
-            debug("Interpret: %s", statement)
-            statement.interpret(self, variables)
 
 #################################
 # "From Selection" Operator
