+++ utils/modules/profile_mk3/__init__.py
@@ -1,0 +1,3 @@
+"""
+This directory contains auxiliary modules used by profile_mk3 node.
+"""
